Move integration tests to the test package
