+++ .gitignore
@@ -1,4 +1,3 @@
 .idea
 **/*.txt
 internal/generated
-test
