+++ internal/server/server.go
@@ -21,7 +21,7 @@ func Run(ctx context.Context, addr string) error {
 	return run(ctx, listener)
 }
 
-func runWithWaitGroup(ctx context.Context, wg *sync.WaitGroup, addr string) error {
+func RunWithWaitGroup(ctx context.Context, wg *sync.WaitGroup, addr string) error {
 	listener, err := net.Listen("tcp4", addr)
 	if err != nil {
 		return err
